+++ lib/dawn/knowledge_base.rb
@@ -258,6 +258,7 @@ require "dawn/kb/osvdb_118954"
 require "dawn/kb/osvdb_119878"
 require "dawn/kb/osvdb_119927"
 require "dawn/kb/osvdb_120415"
+require "dawn/kb/osvdb_120857"
 
 
 
@@ -531,6 +532,7 @@ module Dawn
           Dawn::Kb::OSVDB_119878.new,
           Dawn::Kb::OSVDB_119927.new,
           Dawn::Kb::OSVDB_120415.new,
+          Dawn::Kb::OSVDB_120857.new,
       ]
         # END @cve_security_checks array
         # START @owasp_ror_cheatsheet_checks array