+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -1039,4 +1039,9 @@ describe "The Codesake Dawn knowledge base" do
     sc.should_not   be_nil
     sc.class.should == Dawn::Kb::OSVDB_120415
   end
+  it "must have test for OSVDB_120857" do
+    sc = kb.find("OSVDB_120857")
+    sc.should_not   be_nil
+    sc.class.should == Dawn::Kb::OSVDB_120857
+  end
 end